Move deferred implementation out of `_base`, split into several files.
Move `_base/Deferred` to `Deferred`, `_base/promise` to `promise`, and `tests/_base/Deferred` to `tests/Deferred`. A light-weight `when()` is included under `when`, a more heavy-weight `when()` which returns a promise under `promise/when`.

Changed all references to `Deferred` and `when` to come through `define()`, rather than `dojo.Deferred` and `dojo.when` so modules can make an explicit choice about which to implement.